refactor: extract OSS Insight API client to infrastructure/http
- Add OSS Insight API client to infrastructure/http module
- Update trending.rs to use new API client
- Update trending_unified_view.rs to use new API structure
