+++ src/infrastructure/http/mod.rs
@@ -1,0 +1,3 @@
+pub mod oss_insight_client;
+
+pub use oss_insight_client::OssInsightClient;
