+++ src/infrastructure/mod.rs
@@ -1,6 +1,8 @@
 pub mod cache;
 pub mod config;
 pub mod external;
+pub mod git;
+pub mod http;
 pub mod logging;
 pub mod repository_manager;
 pub mod storage;
